Refactor CpuCore codes
All pieces related to processor including CpuCore are currently in
`processor.hpp`. This separates CpuCore codes into a different file.
